Created a asynch task class that downloads photos asynchronously and got youtube thumbnails downloading
